+++ public/home-page.html
@@ -24,7 +24,7 @@
                 <a href="about.html">About</a>
                 <a href="FQA.html">FAQs</a>
                 <a href="contact-page.html">Contact</a>
-                <a href="signup-form.html">Sign Up</a>
+                <a href="./register">Sign Up</a>
             </div>
         </div>
         <form action="">
@@ -47,7 +47,7 @@
 
                                     <div class="content-section-grid clearfix">
 
-                                        <a href="index.html" class="button nav-link">
+                                        <a href="buy.html" class="button nav-link">
 
                                             <div class="bottom"></div>
 
@@ -88,7 +88,7 @@
 
                                     <div class="content-section-grid clearfix">
 
-                                        <a href="signin-form.html" class="button nav-link">
+                                        <a href="./login" class="button nav-link">
 
                                             <div class="bottom"></div>
